accntchsrpg.js now works every time
Earlier, it only checked for when the account chooser page was done loading. However, it seems that the first time that page is loaded(for example when the account chooser page is not in the user's history), the 'load' event doesn't trigger. So this is a one-size-fits-all rubbish code fix that simply waits for 5 seconds and then sends the message.