Update meetingPage.js
It could be that none of these fixes with meetingPage.js and accountChooserPage.js actually work. I think that is because if for the first time in a long time(need specifics) the meeting or account choosing pages are loaded, then they just don't respond to window.addEventListener('load') and/or setTimeout(). From what I know, we can do without checking if those particular pages have loaded as the extension sees that for itself before mounting the content script in the first place. Or i am wrong again.